requirements.txt
grpcio==1.66.1
protobuf==5.28.0
psycopg2_binary==2.9.9
psycopg2==2.9.9

